Refactor SolrVocabularyInputTerm-SolrVocabularyTerm relationship.
Pull them up together into a common hierarchy instead of having one be
the child of the other. This fixes a bug with JSON serialization of some
objects and better mirrors both the existing and improved post-solrv5.5
relationship between SolrDocument and SolrInputDocument
